refactor importer into a separate component
